+++ seeds/destinationData.js
@@ -1,7 +1,5 @@
 const { Destinations } = require('../models');
 
-const { User } = require('../models');
-
 const destinationdata = [
   {
     city: 'Birmingham',
@@ -205,36 +203,10 @@ const destinationdata = [
   },
 ];
 
-const userdata = [
-  {
-    "name": "Sal",
-    "email": "sal@hotmail.com",
-    "password": "password12345"
-  },
-  {
-    "name": "Lernantino",
-    "email": "lernantino@gmail.com",
-    "password": "password12345"
-  },
-  {
-    "name": "Amiko",
-    "email": "amiko2k20@aol.com",
-    "password": "password12345"
-  },
-  {
-    "name": "Jordan",
-    "email": "jordan99@msn.com",
-    "password": "password12345"
-  },
-  {
-    "name": "Blake",
-    "email": "the_blake@yahoo.com",
-    "password": "password12345"
-  }
-]
+
 
 const seedDestinations = () => Destinations.bulkCreate(destinationdata);
-const seedUsers = () => Users.bulkCreate(userdata);
+
 
 module.exports = seedDestinations;
-module.exports = seedUsers
+
